% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGeoSubmission.R
\name{fastGeoSubmission}
\alias{fastGeoSubmission}
\title{Submit data files to GEO database by sftp in a fast way, using ncft.}
\usage{
fastGeoSubmission(
  fileNames,
  destDir,
  geoFTP = "ftp-private.ncbi.nlm.nih.gov",
  geoPassword = "33\%9uyj_fCh?M16H",
  nPerBatch = 10,
  sleepSecond = 100
)
}
\arguments{
\item{fileNames}{a vector of filenames with full path name to submit}

\item{destDir}{the directory in GEO FTP server. This directory must be created in advance
(in shell):
sftp geo@sftp-private.ncbi.nlm.nih.gov \cr
password: 33%9uyj_fCh?M16H \cr
mkdir destDir \cr
\cr
More detailes can be found in GEO File Transfer Protocol\cr}

\item{geoFTP}{GEO FTP server address. Default is "ftp-private.ncbi.nlm.nih.gov".
This password can be found in GEO File Transfer Protocol.}

\item{geoPassword}{GEO FTP server password. Default is "33%9uyj_fCh?M16H".
This password can be found in GEO File Transfer Protocol.}

\item{nPerBatch}{the number of uploading process in each batch. Default is 10.}

\item{sleepSecond}{the time (in second) to wait after each batch. Default is 100 (seconds).}
}
\description{
The detailed submission process can be found in GEO File Transfer Protocol:
https://www.ncbi.nlm.nih.gov/geo/info/submissionftp.html.
}
\examples{
{
\dontrun{
# Upload files in "./tmp" folder to "ftpDir"
files = dir("./tmp", full.names = T)
fastGeoSubmission(files, "ftpDir")
}
}
}
