% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupplementaryData.R
\name{getSupplementaryData}
\alias{getSupplementaryData}
\title{Download Supplementary file and get the data using GEO accession ID}
\usage{
getSupplementaryData(
  accession = "GSE121213",
  sep = "\\t",
  header = TRUE,
  suppFile = NULL
)
}
\arguments{
\item{accession}{GEO accession ID, the default is "GSE121213", sep = `\\t`, header = TRUE, suppFile = NULL}

\item{sep}{the field separator character. Values on each line of the file are
separated by this character. The default is `\\t`.}

\item{header}{a logical value indicating whether the file contains the names
of the variables as its first line. If missing, the value is determined from
the file format: header is set to TRUE if and only if the first row contains
one fewer field than the number of columns. The default is FALSE.}

\item{suppFile}{supplementary file name. In some of the GEO projects, there are
multiple supplementary files provided. In this case, `getSupplementaryData`
function by default reads the first file and shows a warning. If a supplementary
file name is specified by this parameter, `getSupplementaryData` will read the
data using this file name without showing a warning.
is specified,}
}
\value{
a list, comprised of `data`, file `extension`, `accession`, `filename`,
and `url`.
}
\description{
Download Supplementary file and get the data using GEO accession ID
}
\examples{
\dontrun{
# Get the suplementary data for accession ID: GSE121213
x = getSupplementaryData()
x
x$data

# Get the suplementary data for accession ID: GSE121212
x = getSupplementaryData("GSE121212")
x
x$data

# Get the suplementary data for accession ID: GSE130567
# This suplementary file is a .tar file, which contains 24 .txt.gz files.
x = getSupplementaryData("GSE130567", header = FALSE)
x
x$data

# You will get a warning because of multiple supplementary files
x = getSupplementaryData("GSE69626")
x$data
x$filename

# Specify a file name to avoid the warning
x = getSupplementaryData("GSE69626", suppFile = "GSE69626_RPKMGEO.txt.gz")
x$data
x$filename
}
}
