% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSRRDownloader.R
\name{installSRRDownloader}
\alias{installSRRDownloader}
\title{Instruction to install prerequisite softwares.}
\usage{
installSRRDownloader()
}
\description{
Instruction to install prerequisite softwares.
}
\examples{
{
\\donotrun{
installSRRDownloader()
}
}
}
