% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupplementaryData.R
\name{readTarFile}
\alias{readTarFile}
\title{This is to read the supplementary .tar file in GEO database. In .tar file, there are
usually several .txt.gz (.csv.gz, .tab.gz) files or one .txt.gz (.csv.gz, .tab.gz) file.}
\usage{
readTarFile(tarfile, sep = "\\t", header = FALSE)
}
\arguments{
\item{tarfile}{the name of supplementary .tar file}

\item{sep}{the field separator character. Values on each line of the file are
separated by this character. The default is `\t`.}

\item{header}{a logical value indicating whether the file contains the names
of the variables as its first line. If missing, the value is determined from
the file format: header is set to TRUE if and only if the first row contains
one fewer field than the number of columns. The default is FALSE.}
}
\value{
a list of data frame coming from the .gz files in the .tar file.
}
\description{
This is to read the supplementary .tar file in GEO database. In .tar file, there are
usually several .txt.gz (.csv.gz, .tab.gz) files or one .txt.gz (.csv.gz, .tab.gz) file.
}
