% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSrrID.R
\name{searchSrrID}
\alias{searchSrrID}
\title{Search SRR accession IDs using projectID}
\usage{
searchSrrID(projectID = "PRJNA540657", timeout = 60)
}
\arguments{
\item{projectID}{character, the project accession ID to search in sra database
(https://www.ncbi.nlm.nih.gov/sra). This ID can be searched by
\code{\link{searchProjectID}} function.}

\item{timeout}{numeric, the number of seconds to wait before killing the searching process.}
}
\value{
a SRR documents "runinfo" data frame containing 47 columns, which are
Run, ReleaseDate, LoadDate, spots, bases, spots_with_mates, avgLength,
size_MB, AssemblyName, download_path, Experiment, LibraryName, LibraryStrategy,
LibrarySelection, LibrarySource, LibraryLayout, InsertSize, InsertDev,
Platform, Model, SRAStudy, BioProject, Study_Pubmed_id, ProjectID, Sample,
BioSample, SampleType, TaxID, ScientificName, SampleName, g1k_pop_code, source,
g1k_analysis_group, Subject_ID, Sex, Disease, Tumor, Affection_Status, Analyte_Type,
Histological_Type, Body_Site, CenterName, Submission, dbgap_study_accession,
Consent, RunHash, and ReadHash.\cr
Please note RunHash and ReadHash are not MD5SUM of each file.
}
\description{
Search SRR accession IDs using projectID
}
\examples{
{
\dontrun{
searchSrrID()
searchSrrID("PRJNA543132")
}
}
}
